Change how context struct gets initialized
We had been passing a pointer to the struct to
RegisterServiceCtrlContextEx directly in the initializer of the struct
itself. Code analysis pointed out that the struct is unitialized at that
point, which is technically true, although probably OK because the API
function doesn't use the contents of the struct anyway.

Nonetheless, if we encapsulate the initialization of the handle member
into the constructor, then the struct counts as initialized, even though
it probably looks no different from how it did before this change.

This change opens the door for other enhancements to the context struct.