Use Unicode-aware and Unicode-agnostic logging
Somewhere along the line, messages stopped being logged. I could fix it
by using a non-Unicode logger, which drew attention to how `logger_type`
should be conditionally defined, but the better fix was to use `message`
instead of `smessage`. That should allow the logger to accept either
kind of string.