Change cache key
Use the build-pipeline file as well as the package list because the
package list doesn't change, but we want the cache to get updated to
include the actual packages instead of just the NuGet boilerplate stuff
that's in the cache from before.

Also update the actions/cache version, just because.