Use custom severity level
I'm not sure this is really needed after all. I think it was prompted by
compiler warnings about Boost's `severity_level` enum not being scoped,
but for some reason that warning no longer appears at the first use of
the enum values, so the warning's not really an issue anymore.