Bump setup-msbuild version
This should resolve the error GitHub currently issues about using the
`add-path` command.

I was under the impression that Dependabot would take care of this
automatically because of the pcsz2 3799 pull request. However, that
project didn't get its build dependency updated automatically, either,
so maybe that's not really how it works at all.