Switch to Plog
Google Log appears not to support setting the name of the log file, and
its default name is noisy. It also adds a header to each new file, which
includes the logger's output format, which is wrong whenever a custom
prefix is specified. (That's fixed in mainline, but not released.)

Plog allows us to make fewer changes relative to the original Boost
logging, and allows for configuring things closer to what I want the
logging code and the output to look like.