Replace severity_level with plog::Severity
The Plog time is sufficient, and doesn't require many changes to adopt
in place of the bespoke severity_level. The change results in less
type-casting.

I added a "trace" level to alias `verbose` because I like that name
better, and it reduces the number of changes required elsewhere in the
code.