Add vcpkg.json to solution file
We can store a reference to a schema for this file, but Visual Studio
seems to want to store the absolute path to the vcpkg.json file, which
is unacceptable. Visual Studio doesn't recognize the schema for this
file if I manually change the .sln file to hold a relative path.