Resolve race condition on session registry event
It was possible that a log-on or log-off event would occur while waiting
for registry-change events, and that would affect the make-up of the
session map, so the order used for populating the event list might be
different by the time we use the wait result to determine which
session's event fired.

Now, we disregard the wait result and just look through the list of
sessions for a matching event handle.

Although there's no check for whether we actually found the triggering
event in the list, there shouldn't be any danger of that. The session
map can only have items removed from it in response to the SessionChange
event, which is handled in the service's main thread, the same as the
handler for the session's registry-change event. Any removal of a
session from the map cannot happen while we're also waiting for registry
events.